snapstate: move autorefresh code into autoRefresh helper
This cleans up the snapmgr.go file a bit and moves the
autorefresh logic into a single place/structure which
makes tests easier.

Also adds some more tests around autoRefresh.
