+++ overlord/snapstate/export_test.go
@@ -136,10 +136,19 @@ var (
 
 // refreshes
 var (
+	NewAutoRefresh    = newAutoRefresh
 	NewRefreshHints   = newRefreshHints
 	NewCatalogRefresh = newCatalogRefresh
 )
 
 func MockCatalogRefreshNextRefresh(cr *catalogRefresh, when time.Time) {
 	cr.nextCatalogRefresh = when
+}
+
+func MockRefreshRetryDelay(d time.Duration) func() {
+	origRefreshRetryDelay := refreshRetryDelay
+	refreshRetryDelay = d
+	return func() {
+		refreshRetryDelay = origRefreshRetryDelay
+	}
 }
